handle avalanche_signTransaction
