+++ packages/core-mobile/package.json
@@ -21,20 +21,20 @@
     "gen:glacierApi": "npx openapi-zod-client 'https://glacier-api-dev.avax.network/api-json' -o './app/utils/network/glacierApi.client.ts'"
   },
   "dependencies": {
-    "@avalabs/avalanche-module": "0.1.8",
+    "@avalabs/avalanche-module": "0.0.0-cp-8941-20240815140428",
     "@avalabs/avalanchejs": "4.0.5",
-    "@avalabs/bitcoin-module": "0.1.8",
+    "@avalabs/bitcoin-module": "0.0.0-cp-8941-20240815140428",
     "@avalabs/bridge-unified": "2.1.0",
     "@avalabs/core-bridge-sdk": "3.1.0-alpha.1",
     "@avalabs/core-chains-sdk": "3.1.0-alpha.1",
     "@avalabs/core-coingecko-sdk": "3.1.0-alpha.1",
     "@avalabs/core-utils-sdk": "3.1.0-alpha.1",
     "@avalabs/core-wallets-sdk": "3.1.0-alpha.1",
-    "@avalabs/evm-module": "0.1.8",
+    "@avalabs/evm-module": "0.0.0-cp-8941-20240815140428",
     "@avalabs/glacier-sdk": "3.1.0-alpha.1",
     "@avalabs/k2-mobile": "workspace:*",
     "@avalabs/types": "3.1.0-alpha.1",
-    "@avalabs/vm-module-types": "0.1.8",
+    "@avalabs/vm-module-types": "0.0.0-cp-8941-20240815140428",
     "@blockaid/client": "0.10.0",
     "@coinbase/cbpay-js": "2.2.1",
     "@cubist-labs/cubesigner-sdk": "0.3.28",
